refactor(api): migrate send-invitation-for-admin route in src/team
lastFile:api/src/team/application/certification-center-invitation-controller.js
