+++ api/lib/application/certification-centers/certification-center-controller.js
@@ -154,26 +154,6 @@ const updateReferer = async function (request, h) {
   return h.response().code(204);
 };
 
-const sendInvitationForAdmin = async function (request, h, dependencies = { certificationCenterInvitationSerializer }) {
-  const certificationCenterId = request.params.certificationCenterId;
-  const invitationInformation = await certificationCenterInvitationSerializer.deserializeForAdmin(request.payload);
-
-  const { certificationCenterInvitation, isInvitationCreated } =
-    await usecases.createOrUpdateCertificationCenterInvitationForAdmin({
-      certificationCenterId,
-      email: invitationInformation.email,
-      locale: invitationInformation.language,
-      role: invitationInformation.role,
-    });
-
-  const serializedCertificationCenterInvitation =
-    dependencies.certificationCenterInvitationSerializer.serializeForAdmin(certificationCenterInvitation);
-  if (isInvitationCreated) {
-    return h.response(serializedCertificationCenterInvitation).created();
-  }
-  return h.response(serializedCertificationCenterInvitation);
-};
-
 const certificationCenterController = {
   create,
   createCertificationCenterMembershipByEmail,
@@ -185,7 +165,6 @@ const certificationCenterController = {
   getCertificationCenterDetails,
   getDivisions,
   getStudents,
-  sendInvitationForAdmin,
   update,
   updateReferer,
 };
