+++ api/lib/application/certification-centers/index.js
@@ -170,47 +170,6 @@ const register = async function (server) {
       },
     },
     {
-      method: 'POST',
-      path: '/api/admin/certification-centers/{certificationCenterId}/invitations',
-      config: {
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleCertif,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-                securityPreHandlers.checkAdminMemberHasRoleMetier,
-              ])(request, h),
-            assign: 'hasAuthorizationToAccessAdminScope',
-          },
-        ],
-        handler: certificationCenterController.sendInvitationForAdmin,
-        validate: {
-          params: Joi.object({
-            certificationCenterId: identifiersType.certificationCenterId,
-          }),
-          options: {
-            allowUnknown: true,
-          },
-          payload: Joi.object({
-            data: {
-              attributes: {
-                email: Joi.string().email().required(),
-                language: Joi.string().valid('fr-fr', 'fr', 'en'),
-                role: Joi.string().valid('ADMIN', 'MEMBER').allow(null),
-              },
-            },
-          }),
-        },
-        notes: [
-          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès**\n" +
-            "- Elle permet à un administrateur d'inviter des personnes, déjà utilisateurs de Pix ou non, à être membre d'un centre de certification, via leur **email**",
-        ],
-        tags: ['api', 'invitations', 'certification-center'],
-      },
-    },
-    {
       method: 'GET',
       path: '/api/admin/certification-centers/{certificationCenterId}/invitations',
       config: {
